# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SERVED. On-Demand Process Serving Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs/{id}/location:
    post:
      operationId: publishJobLocation
      tags: [jobs]
      summary: Publish a live GPS ping for the assigned server (server-only)
      description: |
        While a job is `in_progress`, the assigned server may publish their
        current GPS so the requester and attorney can see live progress.
        Caller must be the assigned server. Rejected when job is not in
        `in_progress`.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishLocationBody"
      responses:
        "204":
          description: Ping recorded
        "403":
          description: Not the assigned server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Job is not in_progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /jobs/{id}/location/latest:
    get:
      operationId: getJobLatestLocation
      tags: [jobs]
      summary: Get the latest live GPS ping for an in-progress job
      description: |
        Returns the most recent location ping for the assigned server.
        Visible to the job's requester, the owning attorney, the assigned
        server, and admins. Returns 200 with `location: null` while the job
        is in_progress but no pings have arrived yet, so the client can
        render an empty state. Returns 404 if the job is not currently
        in_progress (no live tracking happens for pending/served/cancelled).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Latest known location (or null while waiting for first ping)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobLatestLocation"
        "404":
          description: Job not found, not visible, or not currently in_progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /jobs/{id}/documents:
    get:
      operationId: listJobDocuments
      tags: [jobs, documents]
      summary: List documents attached to a job (visible to all parties)
      description: |
        Returns the documents linked to this job. Visible to the job's
        requester, any attorney who owns a document on the job, the assigned
        server, and admins. Each document includes a relative `downloadUrl`
        the caller can fetch (the existing /storage/objects/* endpoint
        enforces per-object authorization, so an authenticated browser
        session works for assigned servers and requesters too).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of documents attached to the job
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobDocument"
        "404":
          description: Job not found or not visible to caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /documents:
    get:
      tags: [documents]
      operationId: listDocuments
      summary: List the caller's cloud documents
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
          description: Optional job filter.
      responses:
        '200':
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
    post:
      tags: [documents]
      operationId: createDocument
      summary: Record a newly uploaded document
      description: |
        Call after the file has been PUT to the presigned URL. Persists
        metadata (objectPath, name, size, contentType) and links it to the
        attorney owner (and optional job).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentBody'
      responses:
        '201':
          description: Document recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /documents/{id}:
    delete:
      tags: [documents]
      operationId: deleteDocument
      summary: Delete a cloud document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted.
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Document:
      type: object
      required: [id, name, size, contentType, objectPath, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        size:
          type: integer
          description: Bytes.
        contentType:
          type: string
        objectPath:
          type: string
        jobId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateDocumentBody:
      type: object
      required: [name, size, contentType, objectPath]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
        objectPath:
          type: string
          minLength: 1
        jobId:
          type: integer
...
    PublishLocationBody:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
        lng:
          type: number
          format: double
          minimum: -180
          maximum: 180
        accuracyM:
          type: number
          format: double
          description: Reported GPS accuracy in meters.
        headingDeg:
          type: number
          format: double
          description: Compass heading in degrees clockwise from north.
        speedMps:
          type: number
          format: double
          description: Reported speed in meters per second.
    JobLocationPing:
      type: object
      required: [lat, lng, recordedAt, ageSeconds]
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        accuracyM:
          type: number
          format: double
          nullable: true
        headingDeg:
          type: number
          format: double
          nullable: true
        speedMps:
          type: number
          format: double
          nullable: true
        recordedAt:
          type: string
          format: date-time
        ageSeconds:
          type: integer
          description: Seconds elapsed between recordedAt and the response.
    JobLatestLocation:
      type: object
      required: [jobId, status]
      properties:
        jobId:
          type: integer
        status:
          type: string
          enum: [in_progress, en_route]
        location:
          oneOf:
            - $ref: "#/components/schemas/JobLocationPing"
            - type: "null"
        destination:
          type: object
          nullable: true
          description: Recipient address for map rendering, when geocoded.
          properties:
            address:
              type: string
            lat:
              type: number
              format: double
              nullable: true
            lng:
              type: number
              format: double
              nullable: true
    JobDocument:
      type: object
      required: [id, name, size, contentType, downloadUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        size:
          type: integer
          format: int64
        contentType:
          type: string
        createdAt:
          type: string
          format: date-time
        downloadUrl:
          type: string
          description: Relative URL the browser can fetch (cookie session auth).